Backtrack a minimum of one character when footnote does not match
When a footnote fails to match in Footnote.match_reference,
`lines._index` is moved backwards by a number of characters equal to the
number of newlines remaining in `lines`. This causes buggy behavior when
the input markdown does not contain any newlines, because `lines._index`
does not change.

Instead, always backtrack at least one character, so the token start `[`
is covered in case there are zero newlines.