Ensure LaTeX renderer uses valid \verb delimiter (#149)
The LaTeX renderer uses `\verb` for inline code, but the delimiter is
always a vertical bar, which produces incorrect output when the inline
code also contains a vertical bar (e.g., `example | pipe`).

Rather than using a single static character (i.e., a vertical bar),
this change modifies render_inline_code to search for a non-letter
delimiter that does not appear in the inline code. If no such
delimiter can be found, a RuntimeError is raised to avoid incorrect
output.

Note that the list of possible delimiters is not exhaustive. For
example, numbers (0, 1, 2, etc.) are all valid delimiters for \verb
but are omitted from the search.

Fixes #149