Try all possible delimiters for LaTeX \verb (#149)
This change uses string.punctuation and string.digits to exhaustively
search for delimiters suitable for inline code -- i.e., a delimiter
that does not appear in the inline code. To preserve existing
behavior, the vertical bar (`|`) is always used first followed by
other delimiters that are commonly used (`!`, `"`, `'`, and `=`) [1].

Some delimiters may look weird (e.g., a brace as in `\verb{a|b{`,
which is technically valid) [2], but using these delimiters is
preferable to raising an error when no valid delimiter is found (even
though one might exist). Users may override the default set of
delimiters by specifying the renderer's `verb_delimiters` manually,
which allows braces, brackets, etc. to be omitted.

References #149

[1] https://tex.stackexchange.com/q/203585
[2] https://tex.stackexchange.com/a/387691