Fix for #132, part 1. A link reference definition (footnote) cannot interrupt a paragraph.
The issue was that after parsing a footnote, the parser would keep looking for opening brackets in the same block of lines.
When it found one, it would start parsing it, regardless of the content in-between the first and second footnotes.
Solved by checking that the content in-between two footnotes consists only of white space.