Fix for #89, List Items separated by tab character not parsed correctly.
Also fixes failing examples 312 and 313 in the CommonMark 0.30 spec, due to the way leading space
is now checked for list items.

The direct cause of the reported bug was that only spaces and not tabs were considered
valid separators for list item markers. Another problem was that the implemented tab expansion,
where tabs were always expanded to four spaces, did not work according to the spec, which states
that tabs should be expanded to the nearest tab stop (of width 4).

This fix uses `expandtabs()` to implement the tab stops correctly and moves extraction
of content into the `parse_marker()` and `parse_continuation()` methods. This lets
us implement use cases like "list interrupts a paragraph" and "list item continuation"
in a less error-prone way.