fix: Make filtering children by class in traverse() actually work
The `traverse()` utility function was previously written to filter
children by class using `issubclass(child, klass):`. But the first
argument to `issubclass()` must be a class so this will always raise an
exception when trying to use the functionality. This patch corrects the
call to `isinstance(child, klass)` and adds a test.