Fixed failing example 28 in the CommonMark 0.30 spec: Invalid html entities. (#108)
The problem was that the length of html entities like &#87654321; was not checked,
leading to the generation of invalid output.
The solution was to modify the regex used to match charrefs to include limits on the lengths (as proposed by @pbodnar).
The regex workaround was also moved from the HTMLRenderer class to span_token.tokenize() to make it available to all renderers.