Fix for #89, List Items separated by tab character not parsed correctly. (#164)
Also fixes failing examples 312 and 313 in the CommonMark 0.30 spec due to the way leading space is now checked for list items.

The 1st problem was that `# check if next_line starts List` considered only markers followed
by a space and not by a tab. Another problem was that the implemented tab expansion
to tab stops (of width 4 as per spec) could work only in a limited scope.

This fix uses `expandtabs()` to implement the tab stops correctly and moves extraction
of content into the `parse_marker()` and `parse_continuation()` methods. This lets
us implement use cases like "list interrupts a paragraph" and "list item continuation"
in a less error-prone way.