feat: optional html quotes escaping for HTMLRenderer (#176)
Introduce additional parameters for the HTMLRenderer constructor that
control double and single quotes escaping.
Also revert to previous behavior when single quotes are not escaped (by default).

Not giving the options to renderer classes that inherit from HTMLRenderer - this will be in another commit.

Add the "parameterized" package to write parameterized tests easily.