feat: allow just 1 dash in Table delimiter cells (#131)
In or around [#371](github/markup#371), GFM relaxed constraints on the minimal number of required dashes in table delimiter row.

So we relax the constraints as well. To "compensate" for this change,
we check the format of the delimiter row more thoroughly (although
we still do *not* check number of cells in the delimiter row).