Add types to __init__.py's markdown function (#251)
pyright returns an error when there are no type annotations if using any other `renderer` besides `HtmlRenderer`, because it assumes the type is `type[HtmlRenderer]` instead of `type[BaseRenderer]` like it should be.

To stay compatible with Python 3.5, I've used things from the `typing` module (also used elsewhere in the codebase).