fix: escape early from Table read (#254)
Reading all lines with '|' firstly and only then checking
if the second line is a delimiter row was inefficient,
mainly when combined with activation of check_interrupts_paragraph().

Therefore, we now firstly check the delimiter row
before reading the table body.

Also, add checking of the header row to check_interrupts_paragraph(),
to make the implementation more correct.